C2/w3 done!

+++ Course2/w3/miniproject_3/src/main/java/edu/coursera/concurrent/SieveSequential.java
@@ -36,7 +36,7 @@ public final class SieveSequential extends Sieve {
         final int s = primesList.size();
         for (int i = 0; i < s; ++i) {
             final Integer loopPrime = primesList.get(i);
-            if (candidate % loopPrime.intValue() == 0) {
+            if (candidate % loopPrime == 0) {
                 isPrime = false;
                 break;
             }
